+++ src/main/java/lotto/domain/lotto/Lottos.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.constants.ErrorMessage;
 import lotto.constants.LottoRule;
